requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0

# Web Scraping
requests>=2.26.0
beautifulsoup4>=4.10.0
selenium>=4.1.0
lxml>=4.6.0

# Configuration
pyyaml>=5.4.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Development (optional)
jupyter>=1.0.0
ipykernel>=6.4.0
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.9b0
pylint>=2.11.0

# Jupyter Notebook support
notebook>=6.4.0